cpu/native/periph_timer: make clock agnostic to CPU load
This changes the POSIX clock backing `timer_read()` to
`CLOCK_PROCESS_CPUTIME_ID` in the hope that this make tests in the CI
more repeatable, regardless of background CPU load.

For timeouts still `CLOCK_MONOTONIC` is used, as
`CLOCK_PROCESS_CPUTIME_ID` will not advance when no (RIOT) threads are
runnable and a timeout would never occur.

To do some impedance matching, bookkeeping is added to the time spent
sleeping. This is added on top of the time got from
`CLOCK_PROCESS_CPUTIME_ID`.

**WARNING**: This makes the test even more flaky then before. Either I
mess something up in my implementation, or maybe
`CLOCK_PROCESS_CPUTIME_ID` is just too coarse grained to be of use for
us.