makefiles/docker.inc.mk: Pin riotbuild version with BUILD_IN_DOCKER=1
We had a number of issues being reported that in the end were caused
by building with a version of riotbuild incompatible with the source.

This commits adds a check that the used image is in the listed of
known to work versions, unless `DOCKER_IMAGE` is not the default.
Users overwriting `DOCKER_IMAGE` are expected to know what they are
doing.

This also changes behavior in case no docker image is present: It will
error and tell the use to use docker pull to get the image (and how
to get it).