drivers/periph_timer: add periph_timer_capture API
This adds an API to use timers in capture mode.

A brief look in a number of data sheets has revealed that even primitive
timers such as ATmega or MSP430 timers allow using both capture and
compare mode to be used in parallel. Most use the channel compare
target register as capture register, when a channel is in capture
mode. Some (e.g. ATmega) have a dedicated capture register that
is not shared with compare channels.

The API intents to reflect this by allowing capture channels to
overlap with compare channels, or to have distinct numbers. The number
of the first and the last capture channel for each timer can be
queried by the API. A default (weak symbol) implementation can
be used when all compare channels can be used as capture channels,
which seems to be the case for most timer peripherals.

An implementation is expected to rely on a board or MCU level
configuration that maps channels to event sources (such as internal
triggers or GPIO pins). A function to query which GPIO pin a specific
channel on a given timer refers to is provided, which will return
`GPIO_UNDEF` for internal triggers. Managing internal triggers are out
of scope for a hardware abstracting peripheral API, as those can widely
vary between MCUs. However, managing the capture channels linked to
internal triggers will work with this API in the same way as capture
channels linked to external triggers.

The internal triggers will be particularly useful to implement a
mechanism to synchronize timer peripherals, e.g. to convert timestamps
captured on peripheral A to timestamps on timer peripheral B that is
used in compare mode.