tests/periph/pdm: Initial test
So many things were found and fixed...
I changed the test application so we could verify sample rates.
I added tests that check the values of the sample rates and some sanity check of the actual samples
Added a few changes to the API to make it more testable, such as the variable sample rate and checking the real sample rate.
Implemented variable sample rate.
Cleaned up the python application to only use the output files and removed unneeded dependencies.
Updated the documentation...