doc/CONTRIBUTION.md: document mediation process
As agreed upon in the weekly coordination meeting we want to set up
a channel / tool / process to aid contributors to solve conflicts.

This is especially aimed to empower contributors to solve conflicts
with maintainers, as we were unsure that the existing code of conduct
violation mechanism is sufficient to address conflicts that do not
involve a CoC violation.

Co-authored-by: mguetschow <mikolai.guetschow@tu-dresden.de>