sys/net/gnrc: add dependencies for sock_aux_timestamp
We need `gnrc_tx_sync` to wait for the transmission to complete to
fetch auxiliary data from sending in the upper layers, as otherwise
we have no clue when the data is actually valid.

In addition, we need `netdev_tx_info_timestamp` to tell the netdevs
to provide said timestamp.