sys/psa_crypto: Replace old SHA256 HMAC implemenation with generic HMAC
The old HMAC implementation only supported the SHA256 hashing algorithm
and only implemented the single-part MAC function.
Replace it with the generic HMAC implementation which supports all
hashing algorithms and is already used for the multi-part MAC functions.

A side effect of this commit is that the cryptocell HMAC implementation
is not used anymore. This will be fixed in a later commit which
introduces broad hardare-acceleration for HMAC.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>