core/sched_switch: fix crash with no active thread
The function `sched_switch()` was implemented with the assumption that
there will always be an active thread. This was true until threadless
idle was implemented for Cortex M MCUs: If no thread is runnable and
the running thread exists, there will be no active thread. If from
ISR a thread is then unblocked, `sched_switch()` will be called without
an active thread.

This handles the corner case explicitly when the module
`core_idle_thread` is not used (in other words: for threadless idle).

Fixes https://github.com/RIOT-OS/RIOT/issues/20812