make: export DEVELHELP
This patch exports the make macro `DEVELHELP`. Without this patch, use
of the macro in the following files does not work when the macro is set
in a makefile (such as in Makefile.local or an application's makefile as
demonstrated in dist/Makefile). Inside these files `DEVELHELP` is not
defined under these conditions.

 - pkg/littlefs/Makefile
 - pkg/littlefs2/Makefile
 - sys/stdio_null/Makefile

Note that use of the macro does work in these files when the macro is
set from the command line, without the patch. For example:

``` sh
$make DEVELHELP=1 all
```